Se agregó el date converter

+++ TreeStorage/Element.cpp
@@ -1,8 +1,8 @@
 #include "Element.h"
 #include <iostream>
 
-Element::Element(uint32_t requi, time_t originDate,
-	time_t requestDate, time_t authorizedDate, uint64_t amount, bool hasTax, bool authorized)
+Element::Element(uint16_t requi, time_t originDate,
+	time_t requestDate, time_t authorizedDate, uint32_t amount, bool hasTax, bool authorized)
 	:
 	requi(requi),
 	originDate(originDate),
@@ -20,7 +20,7 @@ void Element::Authorize()
 	authorized = true;
 }
 
-void Element::Print()
+void Element::Print() // not used at the moment
 {
 	std::cout << "Id: " << id << ", Requi: " << requi << ", Monto: $ "
 		<< amount << std::endl;
